Use []Puzzle instead of map[string]Puzzle.
Test failed nondeterministically- I hadn't realized the implicit
assumption that map order would be the same as input order, and
deterministic. In the spirit of not optimizing prematurely, I'm just
going to make this a slice for now, and assert that it prints in the
same order as input... which isn't technically a constraint, but eh.