Update "dimension" to "size".
"size" is the base unit of the grid- e.g. 3 for a (3x3)x(3x3) puzzle.
"dimension" is more accurately the power- e.g. 2 for a two-dimensional
puzzle. So, a normal Sudoku puzzle is (3^2)^2, but you can imagine
Sudoku puzzles in higher dimensions- like, size 2 and dimension 3, where
the numbers 1 to 8 fill 2x2x2 cubes which are in turn arranged in a
4x4x4 cube.