Fix the "complete" function in verify.
D'oh. Can't use 0 as a signal value if it's

This reminds me of one of fmayer's criticisms of Go: there's no Option
(because, of course, no generics). An Option is relatively heavy-weight
versus a signal value... but then you get things like index out of range
panics when you're using a signal value. So there's that.